First Draft
This is an implementation of Xoppa's examples of Model Rendering and
Bullet Physics in LibGDX.
Use arrow keys on PC to move ship around, fire with ctrl, change weapons
with plus or minus, will shoot different shapes wherever you tap on
screen to aim, using the pickRay. Much to be done!